Cleanup and move binding functions
